Remove ticket tailor from default on
Recently ticket tailor have changed thier widget support, and also removed the 'test' event we relied on to populate the widget. The behaviour in local development is for this error to take over the whole screen, which is an irritation when we are trying to make this a friendly OS project.